+++ ai/coach.py
@@ -6,7 +6,7 @@ from multiprocessing.managers import DictProxy
 from time import sleep, time
 from typing import List
 
-from RULEngine.controller import EngineCommand
+from Util.engine_command import EngineCommand
 from Util.team_color_service import TeamColorService
 from ai.executors.debug_executor import DebugExecutor
 from ai.executors.play_executor import PlayExecutor
