+++ ai/executors/play_executor.py
@@ -1,18 +1,17 @@
 # Under MIT License, see LICENSE.txt
-import time
+
 from typing import List, Dict
 import logging
 
 from multiprocessing import Queue
 
-from RULEngine.controller import EngineCommand
+from Util.engine_command import EngineCommand
 from Util import Pose, Position, AICommand, Singleton
 from ai.GameDomainObjects import Player
 from ai.STA.Strategy.human_control import HumanControl
 
-
 from RULEngine.Debug.uidebug_command_factory import UIDebugCommandFactory
-from Util.role import Role
+
 from ai.executors.pathfinder_module import generate_path
 from config.config_service import ConfigService
 from ai.Util.sta_change_command import STAChangeCommand
